Update version num in docs and add 2 options
Some instances of the version number were listed as 0.7.0 even
though the latest version is 0.7.3. These intances were changed to
0.7.3. The change was not made to files in docs/_build/.

One new option was added, and one option was modified to enable an
alternate choice. The new option for the bed subcommand is
--naive-bed-parsing; this enables an input bed file with >=12 columns
to use the simpler assumptions of input bed files with <12 columns.

The new choice for an existing option is for the --chromid option
(for all subcommands implementing that option). Existing choices are
a (as-is), s (short), and l (long). The new choice is n (no-touchy).
As-is is somewhat misleading because it does not actually leave the
input style alone; rather, it attempts to match the output style of
input target style with the input query style on a pair-by-pair
basis. The no-touchy choice is meant to make no changes to the
output chromosome ID style; rather, it will simply keep the exact
chromosome IDs of the input targets, regardless of the style of
the paired input queries. When used, this choice effectively removes
any effect from the update_chromID function.